Add methods to delete a node
The delete method was added to simplify the call to delete a node. The
deleteRec method is triggered from the delete call and will perform the
work of recursively traversing the tree to delete the specified value.
The minValue method was needed to return the inorder successor to a
deleted node with more than one child.